requirements.txt
autogen-core>=0.4.1
autogen-ext>=0.4.1
openai>=1.40.0
python-dotenv>=1.0.1
rich>=13.7.1
